requirements.txt
easydict==1.10
matplotlib==3.5.2
MedPy==0.4.0
numpy==1.21.5
opencv-python==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
tensorboard==2.9.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.2
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.1